Rename SC widget components.
